requirements.txt
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14